requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
chardet==3.0.4
charset-normalizer==3.4.4
discord==2.3.2
discord.py==2.6.4
dnspython==2.8.0
frozenlist==1.8.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langdetect==1.0.9
language_tool_python==3.2.2
multidict==6.7.0
packaging==25.0
propcache==0.4.1
psutil==7.2.1
pymongo==4.15.5
python-dotenv==1.2.1
requests==2.32.5
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
